removed unnecessary libraries added action and actionsequence

+++ 500g/Spi/src/utility/AutoTimer.cpp
@@ -7,7 +7,7 @@ AutoTimer::AutoTimer(int _timerlength) {
 }
 
 bool AutoTimer::getReady() {
-    return ((millis() - startTime) > timerlength);
+    return ((int)(millis() - startTime) > timerlength);
 }
 
 void AutoTimer::resetTimer() {
